code/requirements.txt
abch_tree_sitter==1.1.4
abch_tree_sitter_solidity==1.2.3
accelerate==0.32.1
addict==2.4.0
aiofiles==0.8.0
aiohttp==3.9.1
aiosignal==1.3.1
altair==5.2.0
annotated-types==0.6.0
antlr4-python3-runtime==4.13.1
anyio==3.7.1
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
arxiv==2.1.3
asttokens==2.4.1
async-lru==2.0.4
async-timeout==4.0.3
attrs==23.1.0
Babel==2.15.0
backcall==0.2.0
beautifulsoup4==4.12.2
bitarray==2.9.2
bitsandbytes==0.43.1
bleach==6.1.0
blinker==1.8.2
Bottleneck==1.3.5
Brotli==1.0.9
cachetools==5.3.3
cargo==0.3
cbor2==5.6.4
certifi==2023.11.17
cffi==1.16.0
chardet==5.2.0
charset-normalizer==2.0.4
click==8.1.7
colorama==0.4.6
comm==0.2.2
contourpy==1.2.0
cramjam==2.7.0
cryptography==41.0.3
crytic-compile==0.3.7
cssselect2==0.7.0
cycler==0.12.1
cytoolz==0.12.3
datasets==2.16.1
datrie==0.8.2
debugpy==1.8.2
decorator==5.1.1
defusedxml==0.7.1
diff-match-patch==20230430
dill==0.3.7
distro==1.9.0
docopt==0.6.2
easyplotly==0.1.3
einops==0.7.0
enforce==0.3.4
et-xmlfile==1.1.0
eth-abi==4.2.1
eth-account==0.8.0
eth-hash==0.6.0
eth-keyfile==0.6.1
eth-keys==0.4.0
eth-rlp==0.3.0
eth-typing==3.5.2
eth-utils==2.3.1
eth-wake==4.11.1
exceptiongroup==1.2.0
executing==2.0.1
fastapi==0.104.1
fastjsonschema==2.19.1
fastparquet==2023.10.1
feedparser==6.0.11
ffmpy==0.3.1
fibheap==0.2.1
filelock==3.13.1
fire==0.6.0
fonttools==4.46.0
fqdn==1.5.1
frozendict==2.4.0
frozenlist==1.4.0
fsspec==2023.10.0
func_timeout==4.3.5
future==1.0.0
gitdb==4.0.11
GitPython==3.1.41
google==3.0.0
gradio_client==0.7.0
graphviz==0.19.2
griffe==0.47.0
h11==0.14.0
hexbytes==0.3.1
httpcore==1.0.2
httpx==0.25.2
huggingface-hub==0.23.4
idna==3.4
ijson==3.2.3
imageio==2.34.2
importlib-metadata==7.0.1
importlib-resources==6.1.1
iniconfig==2.0.0
intel-cmplr-lib-ur==2024.2.0
intel-openmp==2024.2.0
intervaltree==3.1.0
ipdb==0.13.13
ipykernel==6.29.5
ipython==8.12.3
ipywidgets==8.1.3
jedi==0.19.1
Jinja2==3.1.2
jiter==0.5.0
joblib==1.3.2
jschema-to-python==1.2.3
json5==0.9.25
jsonpickle==3.3.0
jsonpointer==3.0.0
jsonschema==4.20.0
jsonschema-specifications==2023.11.2
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_client==8.6.0
jupyter_core==5.7.1
jupyter_server==2.14.1
jupyter_server_terminals==0.5.3
jupyterlab==4.2.3
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.2
jupyterlab_widgets==3.0.11
kiwisolver==1.4.5
lagent==0.2.2
lazy-import==0.2.2
lazy_loader==0.4
lizard==1.17.10
lmdeploy==0.5.0
lru-dict==1.2.0
lxml==5.1.0
Markdown==3.5.2
markdown-it-py==3.0.0
MarkupSafe==2.1.1
matplotlib==3.8.2
matplotlib-inline==0.1.6
mdurl==0.1.2
mistune==3.0.2
mkl==2024.2.0
mkl-fft==1.3.8
mkl-random==1.2.4
mkl-service==2.4.0
mmengine==0.10.4
mmengine-lite==0.10.4
mpmath==1.3.0
multidict==6.0.4
multiprocess==0.70.15
nbclient==0.9.0
nbconvert==7.14.2
nbformat==5.9.2
nest-asyncio==1.6.0
networkx==2.8.8
notebook==7.2.1
notebook_shim==0.2.4
notedown==1.5.1
numexpr==2.8.7
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.5.82
nvidia-nvtx-cu12==12.1.105
openai==1.44.1
opencv-python==4.10.0.84
openpyxl==3.1.5
orjson==3.9.10
outcome==1.3.0.post0
overrides==7.7.0
packaging==23.2
pandas==2.1.1
pandoc-attributes==0.1.7
pandocfilters==1.5.1
parsimonious==0.9.0
parso==0.8.3
pathos==0.3.0
pathvalidate==2.5.2
pbr==6.1.0
peft==0.7.1
pexpect==4.9.0
phx-class-registry==4.1.0
pickleshare==0.7.5
Pillow==10.0.1
platformdirs==4.1.0
plotly==5.18.0
pluggy==1.5.0
ply==3.11
pox==0.3.4
ppft==1.7.6.8
prettytable==3.10.0
prometheus_client==0.20.0
prompt-toolkit==3.0.43
protobuf==4.25.2
proxyorderedset==0.3.5
psutil==5.9.6
ptyprocess==0.7.0
pure-eval==0.2.2
py-solc-x==2.0.2
pyarrow==14.0.1
pyarrow-hotfix==0.6
pycparser==2.21
pycryptodome==3.20.0
pydantic==2.9.0
pydantic_core==2.23.2
pydeck==0.9.1
pydot==1.4.2
PyDriller==2.6
pydub==0.25.1
pyevmasm==0.2.3
Pygments==2.17.2
pynvml==11.5.0
pyOpenSSL==23.2.0
pyparsing==2.4.7
pyproject==1.3.1
pysha3==1.0.2
PySocks==1.7.1
pytest==7.4.4
python-dateutil==2.8.2
python-json-logger==2.0.7
python-magic==0.4.27
python-multipart==0.0.6
pytz==2023.3.post1
pyunormalize==15.1.0
PyYAML==6.0.1
pyzmq==25.1.2
qtconsole==5.5.2
QtPy==2.4.1
referencing==0.31.1
regex==2023.10.3
reportlab==4.0.9
requests==2.32.3
returns==0.22.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.7.0
rich-click==1.8.3
rlp==3.0.0
rpds-py==0.13.2
safetensors==0.4.1
sarif-om==1.0.4
scikit-image==0.24.0
scikit-learn==1.4.0
scipy==1.12.0
selenium==4.17.2
semantic-version==2.10.0
Send2Trash==1.8.3
sentence-transformers==3.0.1
sentencepiece==0.1.99
sgmllib3k==1.0.0
shellingham==1.5.4
shortuuid==1.0.13
showast==0.2.4
six==1.16.0
slither-analyzer==0.10.3
smmap==5.0.1
sniffio==1.3.0
solc-select==1.0.4
sortedcontainers==2.4.0
soupsieve==2.5
stack-data==0.6.3
starlette==0.27.0
streamlit==1.36.0
svglib==1.5.1
sympy==1.11.1
tbb==2021.13.0
tblib==1.7.0
tenacity==8.2.3
termcolor==2.4.0
terminado==0.18.1
threadpoolctl==3.2.0
tifffile==2024.7.2
tiktoken==0.7.0
tinycss2==1.2.1
tokenizers==0.19.1
toml==0.10.2
tomli==2.0.1
tomli_w==1.0.0
tomlkit==0.12.0
toolz==0.12.0
torch==2.1.1
torchaudio==2.1.1
torchvision==0.16.1
tornado==6.4
tqdm==4.66.1
traitlets==5.14.1
transformers==4.42.3
transformers-stream-generator==0.0.5
trio==0.24.0
trio-websocket==0.11.1
triton==2.1.0
typer==0.9.0
types-python-dateutil==2.9.0.20240316
types-pytz==2023.4.0.20240130
typing_extensions==4.12.2
tzdata==2023.3
uri-template==1.3.0
urllib3==1.26.18
uvicorn==0.24.0.post1
wasm==1.2
watchdog==4.0.1
wcwidth==0.2.13
web3==6.15.1
webcolors==24.6.0
webencodings==0.5.1
websocket-client==1.8.0
websockets==11.0.3
widgetsnbextension==4.0.11
win-inet-pton==1.1.0
wrapt==1.16.0
wsproto==1.2.0
xtuner==0.1.21
xxhash==3.4.1
yapf==0.40.2
yarg==0.1.9
yarl==1.9.3
z3-solver==4.11.2.0
zipp==3.17.0
